pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["gbifxdl"]

[tool.setuptools_scm]

[project]
name = "gbifxdl"
dynamic = ["version"]
dependencies = [
  "requests",
  "omegaconf",
  "pandas",
  "python-dwca-reader",
  "tqdm",
  "scikit-learn",
  "mmh3",
  "psutil",
  "asyncio",
  "aiofiles",
  "aiohttp_retry",
  "asyncssh",
  "pyarrow",
  "tqdm",
  "pillow",
  "dask",
  "wget",
  "python-dotenv",
]
authors = [
  {name="Guillaume Mougeot", email="guillaume.mougeot@laposte.net"},
]
description = "GBIF downloader."
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
keywords=['GBIF']
license = { text = "GPLv3" }

[project.optional-dependencies]
gpu = [
    "torch",
]

[project.urls]
"Homepage" = "https://github.com/GuillaumeMougeot/gbifxdl"
"Bug Tracker" = "https://github.com/GuillaumeMougeot/gbifxdl/issues"

[project.scripts]
gbifpost = "gbifxdl.run_post:cli"
gbifpipe = "gbifxdl.run_pipeline:cli"
